+++ nequip/model/__init__.py
@@ -6,6 +6,7 @@ from ._weight_init import (
     initialize_from_state,
     load_model_state,
 )
+from ._gmm import GaussianMixtureModelUncertainty
 from ._pair_potential import PairPotential, PairPotentialTerm
 
 from ._build import model_from_config
@@ -23,6 +24,7 @@ __all__ = [
     uniform_initialize_FCs,
     initialize_from_state,
     load_model_state,
+    GaussianMixtureModelUncertainty,
     model_from_config,
     PairPotential,
     PairPotentialTerm,