+++ nequip/utils/_global_options.py
@@ -86,7 +86,8 @@ def _set_global_options(config, warn_on_override: bool = False) -> None:
     )
     fuser = config.get("_jit_fuser", default_fuser)
     # context manager just restores old fuser afterwards
-    torch.jit.fuser(fuser).__enter__()
+    if torch.cuda.is_available():
+        torch.jit.fuser(fuser).__enter__()
     if warn_on_override and fuser != default_fuser:
         # ^ meh assumption, but better than hardcoding getting the old state
         warnings.warn(