+++ CHANGELOG.md
@@ -19,6 +19,9 @@ Most recent change on the bottom.
 - Error (or warn) on unused options in YAML that likely indicate typos
 - `dataset_*_absmax` statistics option
 - `HDF5Dataset` (#227)
+- `include_file_as_baseline_config` for simple modifications of existing configs
+- `nequip-deploy --using-dataset` to support data-dependent deployment steps
+- Support for Gaussian Mixture Model uncertainty quantification (https://doi.org/10.1063/5.0136574)
 
 ### Changed
 - Always require explicit `seed`